% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate correlation coefficients}
\usage{
calculate_correlations(x, num_num_f, num_cat_f, cat_cat_f)
}
\description{
Calculate correlation coefficients between variables in a \code{data.frame} 
using 3 different functions for 3 different possible pairs of vairables:
\itemize{
\item{numeric - numeric}
\item{numeric - categorical}
\item{categorical - categorical}}
}
