\name{CorrGrapheR-package}
\alias{CorrGrapheR-package}
\alias{CorrGrapheR}
\docType{package}
\title{\packageTitle{CorrGrapheR}}
\description{\packageDescription{CorrGrapheR}}
\section{Package Content}{\packageIndices{CorrGrapheR}}
\author{\packageAuthor{CorrGrapheR}}
\section{Maintainer}{\packageMaintainer{CorrGrapheR}}
\keyword{package}
