% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrgrapher.R
\name{corrgrapher}
\alias{corrgrapher}
\alias{corrgrapher.explainer}
\alias{corrgrapher.default}
\title{Create a CorrGrapheR object}
\usage{
corrgrapher(x, ...)

\method{corrgrapher}{explainer}(
  x,
  cutoff = 0.2,
  method = c("pearson", "kendall", "spearman"),
  values = NULL,
  feature_importance = NULL,
  partial_dependency = NULL,
  feature_importance_opts = NULL,
  partial_dependency_opts = NULL
)

\method{corrgrapher}{default}(
  x,
  cutoff = 0.2,
  method = c("pearson", "kendall", "spearman"),
  values = NULL,
  ...
)
}
\arguments{
\item{x}{an object to be used to select the method, which must satisfy conditions:
\itemize{
\item{if \code{data.frame} (default), columns with type \code{numeric} will be selected and called with \code{\link{cor}}.}
\item{if \code{explainer}, methods \code{\link{ingredients::feature_importance}} and \code{\link{ingredients::partial_dependency}} must not return an error. 
Supply them as arguments(\code{feature_importance} or \code{partial_dependency}) 
or supply options to call the functions inside (\code{feature_importance_opts} and \code{partial_dependency})}
}}

\item{...}{other parameters.}

\item{cutoff}{a number. Corelations below this are treated as \strong{no} corelation. Edges corresponding to them will \strong{not} be included in the graph.}

\item{method}{passed directly to \code{\link{cor}} function.}

\item{values}{(Optional) a \code{data.frame} with information abour size of the nodes, containing columns \code{value} and \code{label} (consistent with colnames of \code{x}). Deafult set to equal for all nodes, or (for \code{explainer}) importance of variables.}

\item{feature_importance}{(Optional) an object of \code{feature importance_explainer} class, created by \code{\link{ingredients::feature_importance}} function. If not supported, calculated inside function.}

\item{partial_dependency}{(Optional) an object of \code{aggregated_profile_explainer} class, created by \code{\link{ingredients::partial_dependency}}}
}
\value{
A \code{corrgrapher} object.
}
\description{
Create a CorrGrapheR object before passing it to \code{plot()}.
}
\examples{
df <- as.data.frame(datasets::Seatbelts)[,1:7] # drop the binary target variable
cgr <- corrgrapher(df)
}
\seealso{
\code{\link{plot.corrgrapher}}
}
