% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corrgrapher.R
\name{create_corrgrapher}
\alias{create_corrgrapher}
\alias{create_corrgrapher.explainer}
\alias{create_corrgrapher.default}
\title{Create a CorrGrapheR object}
\usage{
create_corrgrapher(x, ...)

\method{create_corrgrapher}{explainer}(
  x,
  cutoff = 0.2,
  feature_importance = NULL,
  method = c("pearson", "kendall", "spearman")
)

\method{create_corrgrapher}{default}(
  x,
  cutoff = 0.2,
  method = c("pearson", "kendall", "spearman"),
  sizes = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{data.frame}, in which for all \code{numeric} columns calling \code{\link{cor}} makes sense.}

\item{...}{other parameters.}

\item{cutoff}{a number. Corelations below this are treated as \strong{no} corelation. Edges corresponding to them will \strong{not} be included in the graph.}

\item{method}{passed directly to \code{\link{cor}} function.}
}
\value{
A \code{corrgrapher} object.
}
\description{
Create a CorrGrapheR object before passing it to \code{plot()}.
}
\examples{
df <- as.data.frame(datasets::Seatbelts)[,1:7] # drop the binary target variable
cgr <- create_corrgrapher(df)
}
\seealso{
\code{\link{plot.corrgrapher}}
}
