% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html.R
\name{generate_html}
\alias{generate_html}
\title{Generate HTML report}
\usage{
generate_html(cgr, file = "report.html", rewrite = FALSE, ...)
}
\arguments{
\item{cgr}{An object of \code{corrgrapher} class. See \code{\link{craete_corrgrapher}} function.}

\item{file}{File to write content to; passed directly to \code{\link{htmltools::save_html}}.}

\item{rewrite}{If \code{file} exists, should it be overwritten?}
}
\value{
A file of \code{file} name will be generated with either 1 or 2 elements: 
\itemize{
\item{for \code{explainer}, a graph and partial dependency plot on the side.}
\item{for \code{data.frame}, just a graph.}
}
}
\description{
Create an interactive document in HTML based on \code{corrgrapher} object.
}
