% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print_corrgrapher.R
\name{knit_print.corrgrapher}
\alias{knit_print.corrgrapher}
\title{Knitr S3 methods}
\usage{
knit_print.corrgrapher(x, ...)
}
\arguments{
\item{x}{An object of \code{corrgrapher} class. See \code{\link{corrgrapher}} function.}

\item{...}{Other parameters, passed directly to \code{\link[htmltools]{knit_print.shiny.tag}}}
}
\value{
If \code{x} was created from \code{explainer}, 2 objects will be displayed: graph of correlations
on the left and partial dependency plot on the right.
In other case, just a graph will be displayed.
}
\description{
These methods allow \code{corrgrapher} objects to be displayed nicely in knitr/rmarkdown documents.
}
