% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrgraph.R
\name{plot.corrgrapher}
\alias{plot.corrgrapher}
\title{Visualize correlations in a data frame}
\usage{
\method{plot}{corrgrapher}(x, height = 600, width = 1000, ...)
}
\arguments{
\item{x}{a \code{corrgrapher} object. See \code{\link{create_corrgrapher}}.}

\item{height}{Height}

\item{width}{Width}

\item{...}{other parameters, passed directly to \code{\link{visNetwork}} function (such as \code{main}, \code{submain}, etc)}
}
\value{
A \code{\link{visNetwork}} object; graph. On this graph, the edges are treated as springs. 
  The variables correlated \strong{strongly} (positively or negatively) are \strong{close} to each other, 
  and those not (or weakly) correlated - \strong{far} from each other.
}
\description{
Visualize correlations between variables stored in columns, using previously created \code{corrgrapher} object.
}
\examples{
df <- as.data.frame(datasets::Seatbelts)[,1:7] # drop the binary target variable
cgr <- create_corrgrapher(df)
plot(cgr)
}
\seealso{
\code{\link{create_corrgrapher}}
}
