% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_in_knitr.R
\name{plot_in_knitr}
\alias{plot_in_knitr}
\title{Plot in knitr}
\usage{
plot_in_knitr(cgr, ...)
}
\arguments{
\item{cgr}{An object of \code{corrgrapher} class. See \code{\link{craete_corrgrapher}} function.}
}
\value{
A file of \code{file} name will be generated with either 1 or 2 elements: 
\itemize{
\item{for \code{explainer}, a graph and partial dependency plot on the side.}
\item{for \code{data.frame}, just a graph.}
}
}
\description{
Display \code{corrgrapher} object in knitr document.
}
