% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrgrapher.R
\name{corrgrapher}
\alias{corrgrapher}
\alias{corrgrapher.explainer}
\alias{corrgrapher.default}
\title{Create a CorrGrapheR object}
\usage{
corrgrapher(x, ...)

\method{corrgrapher}{explainer}(
  x,
  cutoff = 0.2,
  values = NULL,
  cor_functions = list(),
  ...,
  feature_importance = NULL,
  partial_dependency = NULL,
  feature_importance_opts = NULL,
  partial_dependency_opts = NULL
)

\method{corrgrapher}{default}(x, cutoff = 0.2, values = NULL, cor_functions = list(), ...)
}
\arguments{
\item{x}{an object to be used to select the method, which must satisfy conditions:
\itemize{
\item{if \code{data.frame} (default), columns with type \code{numeric} will be selected and called with \code{\link{cor}}.}
\item{if \code{explainer}, methods \code{\link[ingredients]{feature_importance}} and \code{\link[ingredients]{partial_dependency}} must not return an error. 
Supply them as arguments(\code{feature_importance} or \code{partial_dependency}) 
or supply options to call the functions inside (\code{feature_importance_opts} and \code{partial_dependency})}
}}

\item{...}{other arguments.}

\item{cutoff}{a number. Corelations below this are treated as \strong{no} corelation. Edges corresponding to them will \strong{not} be included in the graph.}

\item{values}{a \code{data.frame} with information abour size of the nodes, containing columns \code{value} and \code{label} (consistent with colnames of \code{x}). Deafult set to equal for all nodes, or (for \code{explainer}) importance of variables.}

\item{cor_functions}{a named \code{list} o functions to pass to \code{\link{calculate_cors}}. Must contain necessary functions from \code{num_num_f}, \code{num_cat_f} or \code{cat_cat_f}. Must contain also \code{max_cor}}

\item{feature_importance}{an object of \code{feature importance_explainer} class, created by \code{\link[ingredients]{feature_importance}} function. If supported, the argument \code{feature_importance_opts} will be ignored.}

\item{partial_dependency}{an object of \code{aggregated_profile_explainer} class, created by \code{\link[ingredients]{partial_dependency}} function. If supported, the argument \code{partial_dependency_opts} will be ignored.}

\item{feature_importance_opts}{a \code{list} of parameters to pass to \code{\link[ingredients]{feature_importance}} function. This argument will be ignored if \code{feature_importance} is supported.}

\item{partial_dependency_opts}{a \code{list} of parameters to pass to \code{\link[ingredients]{partial_dependency}} function. This argument will be ignored if \code{partial_dependency} is supported.}
}
\value{
A \code{corrgrapher} object, consisting of following fields:
\itemize{
\item{\code{nodes} - a \code{data.frame} to pass as argument \code{nodes} to \code{\link{visNetwork}} function}
\item{\code{edges} - a \code{data.frame} to pass as argument \code{edges} to \code{\link{visNetwork}} function}
\item{\code{pds} (if x was of \code{explainer} class) - a splitted \code{partial_dependency_explainer} object. Each item has information about single variable. Passed to \code{\link[ingredients]{plot.aggregated_profiles_explainer}}}
\item{\code{data} - data used to create the object.}
}
}
\description{
Create a CorrGrapheR object before passing it to \code{plot()}.
}
\examples{
# drop the binary target variable
df <- as.data.frame(datasets::Seatbelts)[,1:7] 
cgr <- corrgrapher(df)
}
\seealso{
\code{\link{plot.corrgrapher}}
}
