% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print_corrgrapher.R
\name{knit_print.corrgrapher}
\alias{knit_print.corrgrapher}
\title{Knitr S3 method}
\usage{
\method{knit_print}{corrgrapher}(x, ...)
}
\arguments{
\item{x}{An object of \code{corrgrapher} class. See \code{\link{corrgrapher}} function.}

\item{...}{Other parameters, passed directly to \code{\link[htmltools]{knit_print.shiny.tag}}}
}
\value{
2 objects will be displayed: graph of correlations on the left and a plot on the right. 
If \code{x} was created from \code{explainer}, the plot will visualize partial dependency
of the currently selected variable.
In other case, the plot will visualize distribution of the variable.
}
\description{
This method allows \code{corrgrapher} objects to be displayed nicely in knitr/rmarkdown documents.
}
