% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrgraph.R
\name{plot_corrgraph}
\alias{plot_corrgraph}
\title{Visualize correlations in a data frame}
\usage{
plot_corrgraph(df, cutoff = 0.15)
}
\arguments{
\item{df}{a \code{data frame}, in which \strong{all} columns are \code{numeric} (calling \code{\link{cor}} makes sense).}

\item{cutoff}{a number. Corelations below this are treated as \strong{no} corelation.}
}
\description{
Visualize correlations in a data frame
}
